refactor!: adjust system abstraction
