+++ src/bin/func/Cargo.toml
@@ -25,5 +25,5 @@ name = "scq_ops"
 path = "src/ops.rs"
 
 [[bin]]
-name = "scq_kinesis"
-path = "src/kinesis.rs"
+name = "scq_source"
+path = "src/source.rs"
