+++ src/driver/src/lib.rs
@@ -32,4 +32,3 @@ pub mod deploy;
 pub mod funcgen;
 pub mod logwatch;
 pub mod monitor;
-pub mod query;
