+++ src/query/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "runtime"
+version = "0.1.0"
+description = "A query module to unify data analytics over streams and batches."
+authors = [ "Gang Liao <gangliao@cs.umd.edu>" ]
+license = "Apache-2.0"
+keywords = [ "streaming", "batch", "data analytics", "sql" ]
+edition = "2018"
+
+[dependencies]
+runtime = { path = "../../src/runtime" }
+
+[lib]
+name = "query"
+path = "src/lib.rs"
+crate-type = [ "lib" ]
