+++ src/query/src/lib.rs
@@ -12,25 +12,20 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#![warn(missing_docs)]
-// Clippy lints, some should be disabled incrementally
-#![allow(
-    clippy::float_cmp,
-    clippy::module_inception,
-    clippy::new_without_default,
-    clippy::ptr_arg,
-    clippy::type_complexity,
-    clippy::wrong_self_convention
-)]
+//! A Query API to associate front-end CLI with back-end function generation and
+//! continuous deployment.
 
-//! The runtime contains the context information needed by the lambda function,
-//! such as execution plan and the next lambda functions, which instructs the
-//! lambda instance to perform the correct operation.
+use runtime::datasource::DataSource;
+use runtime::error::Result;
 
-pub mod context;
-pub mod datasource;
-pub mod error;
-pub mod payload;
-pub mod plan;
+/// A SQL query to pull the desired data.
+pub trait Query {
+    /// Deploy the query to the cloud.
+    fn deploy() -> Result<()>;
+}
 
-pub use payload::Payload;
+pub mod batch;
+pub mod stream;
+
+use batch::BatchQuery;
+use stream::StreamQuery;
